+++ app/src/main/java/fr/xgouchet/rehearsal/voice/app/VoiceServiceListener.kt
@@ -1,6 +1,6 @@
 package fr.xgouchet.rehearsal.voice.app
 
 interface VoiceServiceListener {
-    fun readingCue(cueId: Int)
+    fun readingCue(cueId: Long)
     fun stopped()
 }
